+++ src/prompt_selection/distance.py
@@ -1,0 +1,10 @@
+import numpy 
+
+def distance(): 
+    '''
+    each model and dataset, compute euclidean distance between PCs for model and human completions like-ish:
+    sqrt((pc_1_human - pc_1_model)**2 + (pc_2_human - pc_2_model)**2 + …)
+
+    output = id, model, dataset, distance, prompt
+    '''
+    pass
